+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 =====
 
 *   (feature) Add `SerializedType` doctrine type.
+*   (feature) Add base entity helpers:
+    *    `EntityInterface`, `SortableEntityInterface`
+    *   `IdTrait`, `SortOrderTrait`, `TimestampsTrait`